+++ sample/src/main/java/com/canhub/cropper/sample/CropDemoPreset.kt
@@ -1,5 +1,0 @@
-package com.canhub.cropper.sample
-
-enum class CropDemoPreset {
-    RECT, CIRCULAR, CUSTOMIZED_OVERLAY, MIN_MAX_OVERRIDE, SCALE_CENTER_INSIDE, CUSTOM
-}