+++ sample/src/main/java/com/canhub/cropper/sample/crop_image_view/app/CropImageViewFragment.kt
@@ -72,17 +72,16 @@ internal class CropImageViewFragment :
             OptionsDialogBottomSheet.show(childFragmentManager, options, this)
         }
 
-        // TODO Issue #20
-//        binding.searchImage.setOnClickListener {
-//            context?.let { ctx ->
-//                if (CropImage.isExplicitCameraPermissionRequired(ctx)) {
-//                    requestPermissions(
-//                        arrayOf(Manifest.permission.CAMERA),
-//                        CropImage.CAMERA_CAPTURE_PERMISSIONS_REQUEST_CODE
-//                    )
-//                } else activity?.let { CropImage.startPickImageActivity(it) }
-//            }
-//        }
+        binding.searchImage.setOnClickListener {
+            context?.let { ctx ->
+                if (CropImage.isExplicitCameraPermissionRequired(ctx)) {
+                    requestPermissions(
+                        arrayOf(Manifest.permission.CAMERA),
+                        CropImage.CAMERA_CAPTURE_PERMISSIONS_REQUEST_CODE
+                    )
+                } else context?.let { context -> CropImage.startPickImageActivity(context, this) }
+            }
+        }
 
         binding.reset.setOnClickListener {
             binding.cropImageView.apply {
@@ -157,9 +156,12 @@ internal class CropImageViewFragment :
         binding.cropImageView.setOnCropImageCompleteListener(null)
     }
 
-    override fun onSetImageUriComplete(view: CropImageView, uri: Uri, error: Exception) {
-        Log.e("AIC", "Failed to load image by URI", error)
-        Toast.makeText(activity, "Image load failed: " + error.message, Toast.LENGTH_LONG).show()
+    override fun onSetImageUriComplete(view: CropImageView, uri: Uri, error: Exception?) {
+        if (error != null) {
+            Log.e("AIC", "Failed to load image by URI", error)
+            Toast.makeText(activity, "Image load failed: " + error.message, Toast.LENGTH_LONG)
+                .show()
+        }
     }
 
     override fun onCropImageComplete(view: CropImageView, result: CropResult) {